DatasacksInspector: make a confirm dialog for "Delete All PlayerPrefs" option

+++ datasacks/Assets/Datasack/Core/DatasackInspector.cs
@@ -333,9 +333,14 @@ public partial class Datasack
 			GUI.color = Color.red;
 			if (GUILayout.Button( "DELETE ALL PLAYER PREFS"))
 			{
-				DSM.ResetDictionaryIfRunning();
-				PlayerPrefs.DeleteAll();
-				PlayerPrefs.Save();
+				if (EditorUtility.DisplayDialog( "CONFIRM!",
+					"Confirm DELETE all PlayerPrefs stored settings?\n\n" +
+					"(THIS ACTION CANNOT BE UNDONE!)", "DELETE ALL PREFS", "Cancel"))
+				{
+					DSM.ResetDictionaryIfRunning();
+					PlayerPrefs.DeleteAll();
+					PlayerPrefs.Save();
+				}
 			}
 
 			GUILayout.Space(20);
